DisableTelemetry and DisableStudies should be true
In the enterprise policies, that settings *disable* the behaviour
(as the name suggests), so they should be set to true.

The affected policies did not enforce the setting to be enabled, but
just not enforced them to be disabled.
The "about:config" preferences that are generated for the same profile
should still disable telemetry/studies.

Profiles that were exported as "enterprise policy only" should be updated
to enforce disabling these features.