Add a setting to disable Firefox View
Firefox View is a new feature introduced with Firefox 106, which adds a
permantly visible tab, that cannot be closed with some new tab
functions. This setting allows to disable it via the about:config
setting. An alternative way would be to remove it from toolbar, what the
profile generator (currently) cannot do. (Implements #258)