Change merge function to keep input parameters unmodified
If the second argument is modified, then keys from one setting
are added to the dict of another setting. This leads to extensions
the user has not selected being added to policies.json.
If an element that is not a dictionary or list exists in both `a`
and `b`, the value from a is used.
It does not do a full deep copy, copying only dicts that exist in
both `a` and `b`.